+++ webapp/components/survey/NodeDefDetails/DecimalProps.js
@@ -5,6 +5,7 @@ import { FormItem, Input, NumberFormats } from '@webapp/components/form/Input'
 import * as NodeDef from '@core/survey/nodeDef'
 
 import { State } from './store'
+import NodeDefUnitFormItem from './NodeDefUnitFormItem'
 
 const DecimalProps = (props) => {
   const { state, Actions } = props
@@ -16,15 +17,7 @@ const DecimalProps = (props) => {
 
   return (
     <>
-      {NodeDef.isAnalysis(nodeDef) && (
-        <FormItem label="nodeDefEdit.numericProps.unit">
-          <Input
-            className="node-def-unit-input"
-            value={NodeDef.getUnit(nodeDef)}
-            onChange={(value) => Actions.setProp({ state, key: NodeDef.propKeys.unit, value })}
-          />
-        </FormItem>
-      )}
+      <NodeDefUnitFormItem state={state} Actions={Actions} />
       <FormItem label="nodeDefEdit.decimalProps.maxNumberDecimalDigits">
         <Input
           className="node-def-decimal-digits-input"
