+++ webapp/components/survey/NodeDefDetails/IntegerProps.js
@@ -1,28 +1,11 @@
 import PropTypes from 'prop-types'
 
-import { FormItem, Input } from '@webapp/components/form/Input'
-
-import * as NodeDef from '@core/survey/nodeDef'
-
-import { State } from './store'
+import NodeDefUnitFormItem from './NodeDefUnitFormItem'
 
 const IntegerProps = (props) => {
   const { state, Actions } = props
 
-  const nodeDef = State.getNodeDef(state)
-
-  if (!NodeDef.isAnalysis(nodeDef)) return null
-
-  return (
-    <FormItem label="nodeDefEdit.numericProps.unit">
-      <div style={{ width: '10rem' }}>
-        <Input
-          value={NodeDef.getUnit(nodeDef)}
-          onChange={(value) => Actions.setProp({ state, key: NodeDef.propKeys.unit, value })}
-        />
-      </div>
-    </FormItem>
-  )
+  return <NodeDefUnitFormItem state={state} Actions={Actions} />
 }
 
 IntegerProps.propTypes = {
