+++ webapp/components/survey/NodeDefDetails/NodeDefUnitFormItem.js
@@ -1,0 +1,32 @@
+import PropTypes from 'prop-types'
+
+import { FormItem, Input } from '@webapp/components/form/Input'
+
+import * as NodeDef from '@core/survey/nodeDef'
+
+import { State } from './store'
+
+const NodeDefUnitFormItem = (props) => {
+  const { state, Actions } = props
+
+  const nodeDef = State.getNodeDef(state)
+
+  if (!NodeDef.isAnalysis(nodeDef)) return null
+
+  return (
+    <FormItem label="nodeDefEdit.numericProps.unit">
+      <Input
+        className="node-def-unit-input"
+        onChange={(value) => Actions.setProp({ state, key: NodeDef.propKeys.unit, value })}
+        value={NodeDef.getUnit(nodeDef)}
+      />
+    </FormItem>
+  )
+}
+
+NodeDefUnitFormItem.propTypes = {
+  state: PropTypes.object.isRequired,
+  Actions: PropTypes.object.isRequired,
+}
+
+export default NodeDefUnitFormItem
